frida-project/requirements.txt
frida-tools==12.2.1